erros, e app criados, correções em rotas e instalação do swagger

+++ src/controllers/categoryController.js
@@ -1,11 +1,18 @@
 const Category = require('../models/category');
+const missingValues = require('../middlewares/missing-values');
+const NotFound = require('../middlewares/not-found');
 
 class CategoryController {
     async createCategory(req, res) {
         const name = req.body.name;
 
         try {
+            if (!name) {
+                throw new missingValues({ name }, 'O campo nome é obrigatório.');
+            }
+
             const category = await Category.create({ name });
+            
             return res.status(201).send({ success: true, category });
         } catch (error) {
             return res.status(400).send({ error: error.message });
@@ -15,6 +22,7 @@ class CategoryController {
     async listAll(req, res) {
         try {
             const categories = await Category.findAll();
+
             return res.status(200).send(categories);
         } catch (error) {
             return res.status(400).send({ error: error.message });
@@ -25,7 +33,12 @@ class CategoryController {
         const id = req.params.id;
 
         try {
-            const category = await findById(id);
+            const category = await findById(Number(id));
+
+            if (!category) {
+                throw new NotFound(`Categoria com ID ${id} não encontrada.`);
+            }
+
             return res.status(200).send(category);
         } catch (error) {
             return res.status(400).send({ error: error.message });
@@ -33,11 +46,21 @@ class CategoryController {
     }
 
     async updateCategory(req, res) {
-        const id = req.params;
-        const { name } = req.body;
+        const id = req.params.id;
+        const name = req.body.name;
 
         try {
-            const category = await Category.update(
+            const category = await Category.findByPk(Number(id));
+
+            if (!category) {
+                throw new NotFound(`Categoria com ID ${id} não encontrada.`);
+            }
+
+            if (!name) {
+                throw new missingValues({ name }, 'O campo nome é obrigatório.');
+            }
+
+            await Category.update(
                 { name },
                 {
                     where: {
@@ -52,14 +75,21 @@ class CategoryController {
     }
 
     async deleteCategory(req, res) {
-        const id = req.params;
+        const id = req.params.id;
 
         try {
+            const category = await Category.findByPk(Number(id));
+
+            if (!category) {
+                throw new NotFound(`Categoria com ID ${id} não encontrada.`);
+            }
+
             await Category.destroy({
                 where: {
                     id: Number(id)
                 }
             });
+
             return res.status(200).send({ success: true, message: 'Categoria Deletada' });
         } catch (error) {
             return res.status(400).send({ error: error.message });
